requirement.txt
redis==4.5.4
flask==2.2.3
python-dotenv==1.0.0